Fix "(attach_device) cannot find device" by making sure newly-bound devices are configured
Problem Description:
Due to some unknown reasons, devices that work normally with
physical usb host controllers might appear unconfigured after being
bound to a usbip host; as a result, a usbip host will report a list
of bindable devices without any interfaces to remote clients, and
attempts to attach those devices remotely will fail with error
messages like "(attach_device) cannot find device".

Solution:
Adding a check in the binding process to see if a coming device is
configured; if not and it is configurable, calling
usb_choose_configuration() to pick a default configuration, and
setting the device to that configuration. Note that the way of
setting up the default configuration of a device adopted in this
solution is exactly the same as when a device is being authorised
in drivers/usb/core/hub.c